move checksum utils to internal package
