+++ src/main/kotlin/internal/checksums/CreatePublicationChecksum.kt
@@ -1,40 +1,9 @@
-package dev.adamko.gradle.dev_publish.data
+package dev.adamko.gradle.dev_publish.internal.checksums
 
-import dev.adamko.gradle.dev_publish.utils.checksum
-import java.io.File
 import org.gradle.api.file.ConfigurableFileCollection
 import org.gradle.api.file.DirectoryProperty
 import org.gradle.api.provider.*
 import org.gradle.kotlin.dsl.of
-
-
-internal abstract class LoadPublicationChecksum : ValueSource<String, LoadPublicationChecksum.Parameters> {
-
-  interface Parameters : ValueSourceParameters {
-    val checksumFilename: Property<String>
-    val checksumsStore: DirectoryProperty
-  }
-
-  override fun obtain(): String? {
-    val checksumFilename = parameters.checksumFilename.orNull ?: return null
-    val checksumStore = parameters.checksumsStore.get().asFile
-
-    return checksumStore
-      .resolve(checksumFilename)
-      .takeIf(File::exists)
-      ?.readText()
-  }
-
-  internal companion object {
-    fun ProviderFactory.loadPublicationChecksum(
-      configure: Parameters.() -> Unit,
-    ): Provider<String> =
-      of(LoadPublicationChecksum::class) {
-        parameters(configure)
-      }
-  }
-}
-
 
 internal abstract class CreatePublicationChecksum : ValueSource<String, CreatePublicationChecksum.Parameters> {
 
@@ -50,7 +19,7 @@ internal abstract class CreatePublicationChecksum : ValueSource<String, CreatePu
     val checksum = parameters.artifacts
       .map { artifact ->
         val artifactPath = artifact.relativeTo(projectDir).invariantSeparatorsPath
-        "${artifactPath}${FileChecksumSeparator}${artifact.checksum()}"
+        "${artifactPath}$FileChecksumSeparator${artifact.checksum()}"
       }
       .sorted()
       .joinToString("\n")
